accessibility/requirements.txt
urbanaccess==0.2.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
cycler==0.10.0
Cython==0.29.2
Fiona==1.8.4
geopandas==0.4.0
idna==2.8
kiwisolver==1.0.1
lxml==4.2.5
matplotlib==3.0.2
munch==2.3.2
numexpr==2.6.9
pandana==0.4.1
pyparsing==2.3.0
pyproj==1.9.6
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
Shapely==1.6.4.post2
six==1.12.0
tables==3.4.4
urllib3==1.24.1
utm==0.4.2
